+++ drift/telescope/gmrt.py
@@ -4,7 +4,7 @@ import numpy as np
 from scipy.special import jn
 
 from cosmoutils import coord
-from cylsim import telescope
+from drift.core import telescope
 
 
 def jinc(x):
