+++ scripts/interactive.py
@@ -4,16 +4,16 @@ import shutil
 
 import yaml
 
-from cylsim import mpiutil
+from drift.core import mpiutil
 
-from cylsim import cylinder, gmrt, focalplane, restrictedcylinder
-from cylsim import beamtransfer
+from drift.telescope import cylinder, gmrt, focalplane, restrictedcylinder
+from drift.core import beamtransfer
 
-from cylsim import kltransform, doublekl
-from cylsim import psestimation, psmc
-from cylsim import skymodel
+from drift.core import kltransform, doublekl
+from drift.core import psestimation, psmc
+from drift.core import skymodel
 
-from cylsim import projection
+from drift.core import projection
 
 parser = argparse.ArgumentParser(description='Run a simulation job.')
 parser.add_argument('configfile', type=argparse.FileType('r'), help='The configuration file to use.')
