+++ scripts/mk_klfilter.py
@@ -10,10 +10,10 @@ import healpy
 
 from cosmoutils import hputil
 
-from cylsim import cylinder
-from cylsim import beamtransfer
-from cylsim import kltransform
-from cylsim import mpiutil
+from drift.telescope import cylinder
+from drift.core import beamtransfer
+from drift.core import kltransform
+from drift.core import mpiutil
 from mpi4py import MPI
 
 ## Read arguments in.
