+++ scripts/mk_sky_noise.py
@@ -4,7 +4,7 @@ import numpy as np
 import h5py
 
 from cosmoutils import hputil
-from cylsim import beamtransfer, mpiutil
+from drift.core import beamtransfer, mpiutil
 
 btdir = sys.argv[1]
 nside = int(sys.argv[2])
